nl-api-orchestrator/mcp/embed_tools/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3

